puzzle 46/49 (dead dispShape/dispCell statics for the .rodata pool, 0.5f positions, s8 rot, u16 counts; STRIP_UNUSED)

+++ include/puzzle.h
@@ -134,7 +134,7 @@ extern PieceInfo piece_info[];
 extern "C" {
 PieceData* searchItemPieceData(int id, PieceInfo* tbl);
 u8* searchItemModelData(int id, PieceInfo* tbl);
-int PutInCase(u16 id, int num, int type);
+int PutInCase(u16 id, u16 num, int type);
 }
 
 #endif
